ADD gitignore and basic building capabilities
A heavily commented template for CMakeLists.txt was added from
scratch to test and document my knowledge. The "pico_sdk_import.cmake"
was copied from the SDK directory. A dummy main.c was added. It
builds with no problem.

In the CMakeLists.txt I tried a couple of strategies to build
with no additional arguments (board type, wlan connection data) and
seemed to work fine.