ADD RTC and alarm capabilities, on-board LED
The setting and getting of date time using the RP-2040 internal RTC
was tested succesfully. Also the alarm feature in terms of absolute
time (when a particular value of any of the non-minus-one time fields
match with the current time) was tested using a callback function of
mine.

The on-board LED control was also tested. The board is a Pico W so
the on-board LED doesn't connect to a RP-2040 gpio but to a gpio
on the wifi chip (the CYW43), requiring a different setup.