REFACTORIZATION The first serious clean up
A more organized tree directory was built. Functions were taken
out the main.c, renamed with a prefix to show their functionality
area, and put into their own files.

A template for the WiFi configuration was added with useful comments
for a straight forward customization: wlan_setup_data.cmake.template.

In a similar fashion, another file was added with the default timers
parameters: timers_default_data.cmake.

The CMakeLists.txt file was edited to support all these changes.

The README.md was modestly updated with some building instructions.
Also the hardware document file was removed and its contents (not even
close to something updated or correct) was temporarily moved to the
README.md file for the future.