ADDed more network capabilities: UDP server
The UDP server that receives and echoes bytes is working.
	When a NUL terminated c-string "1" is received, the water relay
	coil is	energized, with "3" it's turned off. The same for the
	lights relay with "2" (ON) and "4" (OFF).

The authentication type was added to the
wlan_setup_data.cmake.template using the macros defined by cyw43:

A BUILD_TIME was generated at the cmake level and pass as a macro to
initialize the RTC. Some new functions were added to manipulate the
chosen format to set the time also via udp client request (very soon)
and NTP (may be).

Improvement of the specifications details (see TODO).

Added as a submodule a repository to build a simple UDP client and a
server to test the controller's ability to receive commands accross
the network.